// PortalAutoservicio/src/app/models/detalle-proceso-edd.ts
export class DetalleProcesoedd {
    UNIIDEVALUACIONVIGENTEDADADM: String;
    NOMBREEVALUACIONVIGENTE: String;
    IDUNIDADADMINISTRATIVA: String;
    NOMBREUNIDADADMINISTRATIVA: String;
    TOTALSERVIDORESPUBLICOS: String;
    TOTALSERVIDORESPUBLICOSEVALUADOS: string;
    TOTALSERVIDORESPUBLICOSFALTANTES: string;
    PORCENTAJEAVANZE: string;   
    IDUNIDAD_ASIGNADA: string
    UNIDAD_ASIGNADA: string 
}